Исправлена ошибка с передачей координат через api-функцию list

+++ src/main/java/net/safinart/app/models/Position.java
@@ -1,13 +1,13 @@
 package net.safinart.app.models;
 
 public class Position {
-    public Long lat;
-    public Long lon;
+    public Float lat;
+    public Float lon;
     
     public static Position createEmpty() {
         var pos = new Position();
-        pos.lat = Long.valueOf(0);
-        pos.lon = Long.valueOf(0);
+        pos.lat = Float.valueOf(0);
+        pos.lon = Float.valueOf(0);
         return pos;
     }
 }
